feat: change styles for Line type
