+++ libs/map/ui-map/src/index.ts
@@ -12,4 +12,4 @@ export { ComparisonToolSlider } from './lib/layers/comparison-mode-image-layer/c
 export { ComparisonToolButton } from './lib/layers/comparison-mode-image-layer/comparison-tool-button/comparison-tool-button.component';
 export { MeasureDistanceLayer } from './lib/layers/measure-distance-layer/measure-distance-layer.component';
 export { MeasureDistanceButton } from './lib/layers/measure-distance-layer/measure-distance-button.component';
-export { MeasureDistancePanel } from './lib/layers/measure-distance-layer/measure-distance-panel.component';
+export { MeasureDistancePanel } from './lib/layers/measure-distance-layer/measure-distance-panel/measure-distance-panel.component';
